+++ src/main/kotlin/com/petqua/exception/auth/AuthException.kt
@@ -1,13 +1,13 @@
-package com.petqua.common.exception.auth
+package com.petqua.exception.auth
 
 import com.petqua.common.exception.BaseException
 import com.petqua.common.exception.BaseExceptionType
 
-class OauthClientException(
-    private val oauthClientExceptionType: OauthClientExceptionType
+class AuthException(
+    private val authExceptionType: AuthExceptionType
 ) : BaseException() {
 
     override fun exceptionType(): BaseExceptionType {
-        return oauthClientExceptionType
+        return authExceptionType
     }
 }
