+++ src/main/kotlin/com/petqua/exception/auth/OauthClientException.kt
@@ -1,13 +1,13 @@
-package com.petqua.common.exception.auth
+package com.petqua.exception.auth
 
 import com.petqua.common.exception.BaseException
 import com.petqua.common.exception.BaseExceptionType
 
-class AuthException(
-    private val authExceptionType: AuthExceptionType
+class OauthClientException(
+    private val oauthClientExceptionType: OauthClientExceptionType
 ) : BaseException() {
 
     override fun exceptionType(): BaseExceptionType {
-        return authExceptionType
+        return oauthClientExceptionType
     }
 }
